requirements.txt
click
numpy
Pillow
tensorflow-gpu
matplotlib
